Feb/07/2013
1 Added covariates option to the HE regression.
So far the regression is linear only.  For case-control data, a logistic regression should be added.
2 Added LittleEndianDataInputStream, written by Zhixiang, into the package
3 Moved H2Transformer to src/he/h2trans

=========================================================================

--qcovar
It specifies the quantitative phenotype file.  The format of the file is family id, individual id, and quantitative covariates.

--qcovar-num
The indexes  for the quantitative covariates.  When this option is not specified, all the quantitative covariates will be employed.  The format to specify the covariates are as below:
--qcovar-num 1,2,3,4,5,6
--qcovar-num 1,2-4,6

--covar
It specifies the quantitative phenotype file.  The format of the file is family id, individual id, and categorical covariates.  For example, if one covariate has three items, say Japan, China, and Korea, three numerical values will be assigned to each.

--covar-num
The indexes  for the categorical covariates.  When this option is not specified, all the categorical covariates will be employed.  The format to specify the covariates are as below:
--covar-num 1,2,3,4,5,6
--covar-num 1,2-4,6